requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
numpy==2.0.2